moodrss/requirements.txt
Django==5.0.6
asgiref==3.8.1
click==8.1.7
feedparser==6.0.11
joblib==1.4.2
nltk==3.8.1
python-dateutil==2.9.0.post0
regex==2024.5.15
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
gunicorn==22.0.0
